+++ test/runtests.jl
@@ -36,10 +36,10 @@ mutable struct Agent6 <: AbstractAgent
   id::Int
   pos::NTuple{2, Float64}
   vel::NTuple{2, Float64}
-  diameter::Float64
+  weight::Float64
 end
 
-@testset "all tests" begin
+@testset "Agents.jl Tests" begin
 
 include("api_tests.jl")
 include("space_test.jl")
@@ -47,5 +47,6 @@ include("interaction_tests.jl")
 include("data_collector_test.jl")
 include("CA_test.jl")
 include("continuousSpace_tests.jl")
+include("collisions_tests.jl")
 
 end