+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * new function `space_neighbors`, which works for any space. It always and consistently returns the **IDs** of neighbors irrespectively
   of the spatial structure.
 * Reworked the public API of `GridSpace` to be simpler: position must be `NTuple{Int}`. As a result `vertex2coord` and stuff no longer exported, since they are obsolete.
+* New convenience function `allagents`
+* New continuous space functions `nearest_neighbor` and `elastic_collision!`
+* New iterator `interacting_pairs`
 
 # v2.1
 * Renamed the old scheduler `as_added` to `by_id`, to reflect reality.